requirements.txt
# SDA-U Framework Implementation Requirements
# 선택적 언러닝과 적응형 도메인 적응을 위한 통합 프레임워크

# 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0

# 수치 계산
numpy>=1.21.0
scipy>=1.7.0

# 머신러닝 유틸리티
scikit-learn>=1.0.0

# 시각화
matplotlib>=3.5.0
seaborn>=0.11.0

# 진행률 표시
tqdm>=4.62.0

# 주피터 노트북 지원
jupyter>=1.0.0
ipykernel>=6.0.0

# 데이터 처리
pandas>=1.3.0

# 이미지 처리
Pillow>=8.3.0

# Office-31 데이터셋 다운로드 (Google Drive)
gdown>=4.6.0

# 선택사항: CUDA 지원 (GPU 사용시 별도 설치 필요)
# GPU 사용자는 아래 명령을 별도로 실행하세요:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# 개발 도구 (선택사항)
# black>=22.0.0
# flake8>=4.0.0
# pytest>=7.0.0